fix: move sql db script into validator folder
