+++ sturdy/validator/sql.py
